# docs/api/dgraph-openapi.yaml
openapi: 3.0.3
info:
  title: DGraph Service API
  description: |
    Off-chain and social data. All social interactions (posts, feed, interactions),
    prediction markets (list, bets, payouts), governance variables, subscription status,
    location preferences and spots, ad campaigns/impressions/analytics, and safety
    run through this service (read/write from Dgraph or in-memory fallback).
  version: 1.0.0
servers:
  - url: http://localhost:3003
    description: Local default
  - url: https://gql.dlux.io
    description: Test/production
...
paths:
...
  /governance/variables/{name}:
    get:
      tags: [governance]
      summary: Get one variable
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Variable
        '404':
          description: Variable not found
...
  /markets:
    post:
      tags: [markets]
      summary: Create prediction market (returns tx id; on-chain creation separate)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                dappId: { type: string }
                safetyMetric: { type: string }
      responses:
        '200':
          description: transactionId
...
  /campaigns:
    get:
      tags: [campaigns]
      summary: List campaigns
      parameters:
        - name: advertiser
        - name: status
      responses:
        '200':
          description: campaigns
    post:
      tags: [campaigns]
      summary: Create campaign
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                advertiser: { type: string }
                title: { type: string }
                targetUrl: { type: string }
                placements: { type: array, items: { type: string } }
                bid: { type: number }
                totalBudget: { type: number }
      responses:
        '201':
          description: campaign
  /campaigns/{id}:
    get:
      tags: [campaigns]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: campaign
    delete:
      tags: [campaigns]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: cancelled
...
  /analytics/campaign/{campaignId}:
    get:
      tags: [campaigns]
      parameters:
        - name: campaignId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: analytics
  /analytics/advertiser/{advertiser}:
    get:
      tags: [campaigns]
      parameters:
        - name: advertiser
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: analytics